@@ -1,6 +1,6 @@
 {
   "name": "mkdocs-material",
-  "version": "7.0.0",
+  "version": "7.0.3",
   "description": "A Material Design theme for MkDocs",
   "keywords": [
     "mkdocs",
@@ -24,14 +24,13 @@
     "url": "https://github.com/squidfunk/mkdocs-material.git"
   },
   "scripts": {
-    "build": "run-s build:*",
-    "build:clean": "rimraf material",
-    "build:tools": "ts-node -T tools/build --optimize",
+    "build": "rimraf material && ts-node -T tools/build --optimize",
+    "build:dirty": "ts-node -T tools/build --dirty",
     "clean": "rimraf material",
-    "lint": "run-p lint:*",
+    "lint": "npm run lint:scss && npm run lint:ts",
     "lint:scss": "stylelint \"src/assets/**/*.scss\"",
     "lint:ts": "eslint --cache \"src/**/*.ts\"",
-    "start": "ts-node -T tools/build --watch"
+    "start": "ts-node -T tools/build --verbose --dirty --watch"
   },
   "dependencies": {
     "clipboard": "^2.0.6",
@@ -40,12 +39,12 @@
     "fuzzaldrin-plus": "^0.6.0",
     "lunr": "^2.3.9",
     "lunr-languages": "^1.4.0",
-    "rxjs": "^7.0.0-beta.10"
+    "rxjs": "^7.0.0-beta.12"
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^5.15.2",
     "@mdi/svg": "^5.9.55",
-    "@primer/octicons": "^12.0.0",
+    "@primer/octicons": "^12.1.0",
     "@types/clipboard": "^2.0.1",
     "@types/escape-html": "1.0.0",
     "@types/fuzzaldrin-plus": "^0.6.1",
@@ -59,8 +58,8 @@
     "autoprefixer": "^10.2.4",
     "chokidar": "^3.5.1",
     "cssnano": "^4.1.10",
-    "esbuild": "^0.8.50",
-    "eslint": "^7.20.0",
+    "esbuild": "^0.8.53",
+    "eslint": "^7.21.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsdoc": "^32.1.0",
@@ -70,7 +69,6 @@
     "html-minifier": "^4.0.0",
     "material-design-color": "^2.3.2",
     "material-shadows": "^3.0.1",
-    "npm-run-all": "^4.1.5",
     "postcss": "^8.2.5",
     "postcss-inline-svg": "^5.0.0",
     "preact": "^10.5.12",
