@@ -30,9 +30,9 @@ import { getElementOrThrow, getElements } from "~/browser"
  * Component
  */
 export type ComponentType =
-  | "icon-search"                      /* Icon search */
-  | "icon-search-query"                /* Icon search input */
-  | "icon-search-result"               /* Icon search results */
+  | "iconsearch"                       /* Icon search */
+  | "iconsearch-query"                 /* Icon search input */
+  | "iconsearch-result"                /* Icon search results */
 
 /**
  * A component
@@ -49,37 +49,50 @@ export type Component<
   }
 
 /* ----------------------------------------------------------------------------
+ * Helper types
+ * ------------------------------------------------------------------------- */
+
+/**
+ * Component type map
+ */
+interface ComponentTypeMap {
+  "iconsearch": HTMLElement            /* Icon search */
+  "iconsearch-query": HTMLInputElement /* Icon search input */
+  "iconsearch-result": HTMLElement     /* Icon search results */
+}
+
+/* ----------------------------------------------------------------------------
  * Functions
  * ------------------------------------------------------------------------- */
 
 /**
  * Retrieve the element for a given component or throw a reference error
  *
- * @template T - Element type
+ * @template T - Component type
  *
  * @param type - Component type
  * @param node - Node of reference
  *
  * @returns Element
  */
-export function getComponentElement<T extends HTMLElement>(
-  type: ComponentType, node: ParentNode = document
-): T {
+export function getComponentElement<T extends ComponentType>(
+  type: T, node: ParentNode = document
+): ComponentTypeMap[T] {
   return getElementOrThrow(`[data-mdx-component=${type}]`, node)
 }
 
 /**
  * Retrieve all elements for a given component
  *
- * @template T - Element type
+ * @template T - Component type
  *
  * @param type - Component type
  * @param node - Node of reference
  *
  * @returns Elements
  */
-export function getComponentElements<T extends HTMLElement>(
-  type: ComponentType, node: ParentNode = document
-): T[] {
+export function getComponentElements<T extends ComponentType>(
+  type: T, node: ParentNode = document
+): ComponentTypeMap[T][] {
   return getElements(`[data-mdx-component=${type}]`, node)
 }
