@@ -20,77 +20,77 @@
  * IN THE SOFTWARE.
  */
 
-import { Observable, merge } from "rxjs"
-
-import { configuration } from "~/_"
-import { requestJSON } from "~/browser"
-
-import { Component, getComponentElement } from "../../_"
 import {
-  IconSearchQuery,
-  mountIconSearchQuery
-} from "../query"
+  Observable,
+  combineLatest,
+  fromEvent,
+  merge
+} from "rxjs"
 import {
-  IconSearchResult,
-  mountIconSearchResult
-} from "../result"
+  delay,
+  distinctUntilChanged,
+  filter,
+  map,
+  withLatestFrom
+} from "rxjs/operators"
+
+import { watchElementFocus } from "~/browser"
+
+import { Component } from "../../_"
 
 /* ----------------------------------------------------------------------------
  * Types
  * ------------------------------------------------------------------------- */
 
 /**
- * Icon category
+ * Icon search query
  */
-export interface IconCategory {
-  base: string                         /* Category base URL */
-  data: Record<string, string>         /* Category data */
+export interface IconSearchQuery {
+  value: string                        /* Query value */
+  focus: boolean                       /* Query focus */
 }
-
-/**
- * Icon search index
- */
-export interface IconSearchIndex {
-  icons: IconCategory                  /* Icons */
-  emojis: IconCategory                 /* Emojis */
-}
-
-/* ------------------------------------------------------------------------- */
-
-/**
- * Icon search
- */
-export type IconSearch =
-  | IconSearchQuery
-  | IconSearchResult
 
 /* ----------------------------------------------------------------------------
  * Functions
  * ------------------------------------------------------------------------- */
 
 /**
- * Mount icon search
+ * Mount icon search query
  *
- * @param el - Icon search element
+ * @param el - Icon search query element
  *
- * @returns Icon search component observable
+ * @returns Icon search query component observable
  */
-export function mountIconSearch(
-  el: HTMLElement
-): Observable<Component<IconSearch>> {
-  const config = configuration()
-  const index$ = requestJSON<IconSearchIndex>(
-    `${config.base}/overrides/assets/javascripts/icon_search_index.json`
-  )
+export function mountIconSearchQuery(
+  el: HTMLInputElement
+): Observable<Component<IconSearchQuery, HTMLInputElement>> {
 
-  /* Retrieve nested components */
-  const query  = getComponentElement("icon-search-query", el)
-  const result = getComponentElement("icon-search-result", el)
-
-  /* Create and return component */
-  const query$ = mountIconSearchQuery(query as HTMLInputElement)
-  return merge(
-    query$,
-    mountIconSearchResult(result, { index$, query$ })
+  /* Intercept focus and input events */
+  const focus$ = watchElementFocus(el)
+  const value$ = merge(
+    fromEvent(el, "keyup"),
+    fromEvent(el, "focus").pipe(delay(1))
   )
+    .pipe(
+      map(() => el.value),
+      distinctUntilChanged()
+    )
+
+  /* Log search on blur */
+  focus$
+    .pipe(
+      filter(active => !active),
+      withLatestFrom(value$)
+    )
+      .subscribe(([, value]) => {
+        const path = document.location.pathname
+        if (value.length)
+          ga("send", "pageview", `${path}?q=[icon]+${value}`)
+      })
+
+  /* Combine into single observable */
+  return combineLatest([value$, focus$])
+    .pipe(
+      map(([value, focus]) => ({ ref: el, value, focus })),
+    )
 }
